requirements.txt
blinker==1.9.0
cachelib==0.13.0
click==8.3.1
et_xmlfile==2.0.0
Flask==3.1.3
Flask-Session==0.8.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msgspec==0.20.0
numpy==2.4.2
openpyxl==3.1.5
pandas==3.0.1
pygame==2.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
six==1.17.0
Werkzeug==3.1.6
